Refactor and cleanup

+++ include/bvh.h
@@ -2,13 +2,13 @@
 #define RAYTRACER_BVH_H
 
 #include <iostream>
-#import "object.h"
+#include "object.h"
 
 class BVH : public Object {
 public:
     BVH() {}
 
-    BVH(Object **surfaces, int num_surfaces);
+    BVH(Object **surfaces, size_t num_surfaces);
 
     bool intersects(const Ray &ray, float tmin, float tmax, HitRecord &hit_record, bool backfaceCulling) override;
